feat(auth): implement JWT token management and authentication middleware
- Introduced JWT authentication middleware to handle token validation, including checks for revoked tokens.
- Added support for access and refresh tokens, enhancing the security of user sessions.
- Updated the user handler and repository to integrate new authentication features, including user login and registration processes.
- Removed deprecated OAuth and passkey handling methods from the user handler to streamline authentication logic.
- Enhanced CORS middleware to specify allowed headers for improved security.
