+++ internal/model/auth/scope.go
@@ -1,8 +1,16 @@
 package model
 
+// Token 类型常量，对应 JWT claims 中的 "typ" 字段。
+//
+//   - session：浏览器用户会话的 access token（短期，前端存 JS 内存）
+//   - access：管理面板签发的 API token（长期，面向 CLI/集成/MCP）
+//   - refresh：静默刷新专用（长期，存 HttpOnly Cookie）
+//
+// ParseToken() 仅接受 session/access，ParseRefreshToken() 仅接受 refresh，两者互不混用。
 const (
 	TokenTypeSession = "session"
 	TokenTypeAccess  = "access"
+	TokenTypeRefresh = "refresh"
 )
 
 const (
