+++ internal/service/auth/oauth_adapter.go
@@ -1,4 +1,4 @@
-package service
+package auth
 
 import (
 	"errors"
