+++ internal/service/auth/oauth_service_test.go
@@ -1,4 +1,4 @@
-package service
+package auth
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ func TestParseAndValidateClientRedirect_Allowed(t *testing.T) {
 	cfg := config.Config()
 	cfg.Auth.Redirect.AllowedReturnURLs = []string{"https://app.example.com/auth"}
 
-	svc := &UserService{}
+	svc := &AuthService{}
 	u, err := svc.parseAndValidateClientRedirect("https://app.example.com/auth?from=test")
 	if err != nil {
 		t.Fatalf("expected allow redirect, got err: %v", err)
@@ -24,7 +24,7 @@ func TestParseAndValidateClientRedirect_Denied(t *testing.T) {
 	cfg := config.Config()
 	cfg.Auth.Redirect.AllowedReturnURLs = []string{"https://app.example.com/auth"}
 
-	svc := &UserService{}
+	svc := &AuthService{}
 	_, err := svc.parseAndValidateClientRedirect("https://evil.example.net/auth")
 	if err == nil {
 		t.Fatalf("expected deny redirect")
