+++ internal/service/auth/provider.go
@@ -1,0 +1,8 @@
+package auth
+
+import "github.com/google/wire"
+
+var ProviderSet = wire.NewSet(
+	NewAuthService,
+	wire.Bind(new(Service), new(*AuthService)),
+)
