+++ internal/service/user/ports.go
@@ -2,8 +2,6 @@ package service
 
 import (
 	"context"
-	"encoding/json"
-	"time"
 
 	authModel "github.com/lin-snow/ech0/internal/model/auth"
 	model "github.com/lin-snow/ech0/internal/model/user"
@@ -12,7 +10,6 @@ import (
 )
 
 type Service interface {
-	Login(loginDto *authModel.LoginDto) (string, error)
 	InitOwner(registerDto *authModel.RegisterDto) error
 	Register(registerDto *authModel.RegisterDto) error
 	UpdateUser(ctx context.Context, userdto model.UserInfoDto) error
@@ -21,17 +18,6 @@ type Service interface {
 	GetOwner() (model.User, error)
 	DeleteUser(ctx context.Context, id string) error
 	GetUserByID(userId string) (model.User, error)
-	BindOAuth(ctx context.Context, provider string, redirectURI string) (string, error)
-	GetOAuthLoginURL(provider string, redirectURI string) (string, error)
-	HandleOAuthCallback(provider string, code string, state string) (string, error)
-	GetOAuthInfo(ctx context.Context, provider string) (model.OAuthInfoDto, error)
-	PasskeyRegisterBegin(ctx context.Context, rpID, origin, deviceName string) (authModel.PasskeyRegisterBeginResp, error)
-	PasskeyRegisterFinish(ctx context.Context, rpID, origin, nonce string, credential json.RawMessage) error
-	PasskeyLoginBegin(rpID, origin string) (authModel.PasskeyLoginBeginResp, error)
-	PasskeyLoginFinish(rpID, origin, nonce string, credential json.RawMessage) (string, error)
-	ListPasskeys(ctx context.Context) ([]authModel.PasskeyDeviceDto, error)
-	DeletePasskey(ctx context.Context, passkeyID string) error
-	UpdatePasskeyDeviceName(ctx context.Context, passkeyID string, deviceName string) error
 }
 
 type (
@@ -49,29 +35,6 @@ type UserRepo interface {
 	DeleteUser(ctx context.Context, id string) error
 }
 
-type IdentityRepo interface {
-	BindOAuth(ctx context.Context, userID string, provider, oauthID, issuer, authType string) error
-	GetUserByOAuthID(ctx context.Context, provider, oauthID string) (model.User, error)
-	GetUserByOIDC(ctx context.Context, provider, oauthID, issuer string) (model.User, error)
-	GetOAuthInfo(userId string, provider string) (model.UserExternalIdentity, error)
-	GetOAuthOIDCInfo(userId string, provider string, issuer string) (model.UserExternalIdentity, error)
-}
-
-type PasskeyRepo interface {
-	CreatePasskey(ctx context.Context, passkey *authModel.Passkey) error
-	ListPasskeysByUserID(userID string) ([]authModel.Passkey, error)
-	GetPasskeyByCredentialID(credentialID string) (authModel.Passkey, error)
-	UpdatePasskeyUsage(ctx context.Context, passkeyID string, signCount uint32, lastUsedAt int64) error
-	UpdatePasskeyDeviceName(ctx context.Context, userID, passkeyID string, deviceName string) error
-	DeletePasskeyByID(ctx context.Context, userID, passkeyID string) error
-}
-
-type ChallengeStore interface {
-	CacheSetPasskeySession(key string, val any, ttl time.Duration)
-	CacheGetPasskeySession(key string) (any, error)
-	CacheDeletePasskeySession(key string)
-}
-
 type InstallStateRepo interface {
 	IsInitialized(ctx context.Context) (bool, error)
 	MarkInitialized(ctx context.Context) error
@@ -79,8 +42,5 @@ type InstallStateRepo interface {
 
 type Repository interface {
 	UserRepo
-	IdentityRepo
-	PasskeyRepo
-	ChallengeStore
 	InstallStateRepo
 }
