+++ web/src/lib/file/api/adapter.ts
@@ -8,6 +8,7 @@ import {
   fetchUploadFile,
 } from '@/service/api'
 import { getApiUrl } from '@/service/request/shared'
+import { useAuthStore } from '@/stores/auth'
 import type {
   ExternalFileInput,
   FileEntity,
@@ -92,16 +93,8 @@ export async function updateFileMeta(input: UpdateFileMetaInput): Promise<FileEn
   return normalizeFileEntity(data)
 }
 
-function readAuthTokenFromStorage(): string {
-  if (typeof window === 'undefined') return ''
-  const raw = window.localStorage.getItem('token') || ''
-  const token = raw.replace(/^"|"$/g, '').trim()
-  if (!token || token === 'null' || token === 'undefined') return ''
-  return token
-}
-
 export function buildStreamUrl(fileId: string, t = Date.now()): string {
-  const token = readAuthTokenFromStorage()
+  const token = useAuthStore().accessToken
   const tokenQuery = token ? `&token=${encodeURIComponent(token)}` : ''
   return `${getApiUrl()}/file/${fileId}/stream?t=${t}${tokenQuery}`
 }
